SHLVL is different according to each distro

+++ libft/ft_itoa.c
@@ -3,19 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   ft_itoa.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: alessandro <alessandro@student.42.fr>      +#+  +:+       +#+        */
+/*   By: adorigo <adorigo@student.s19.be>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/10 12:09:03 by adorigo           #+#    #+#             */
-/*   Updated: 2020/11/05 12:31:41 by alessandro       ###   ########.fr       */
+/*   Updated: 2020/11/25 14:01:49 by adorigo          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static int		count_len(int nbr)
+static int		count_len(long nbr)
 {
 	int				count;
-	unsigned int	unbr;
+	unsigned long	unbr;
 
 	if (nbr == 0)
 		return (1);
@@ -34,7 +34,7 @@ static int		count_len(int nbr)
 	return (count);
 }
 
-char			*ft_itoa(long n)
+char			*ft_itoa(unsigned long n)
 {
 	char			*str;
 	int				len;
@@ -46,7 +46,7 @@ char			*ft_itoa(long n)
 		return (NULL);
 	str[len] = '\0';
 	is_negative = 0;
-	unbr = n;
+	unbr = (unsigned long)n;
 	if (n < 0)
 	{
 		is_negative = 1;
